+++ src/routes/project/mod.rs
@@ -8,6 +8,7 @@ pub mod get;
 pub mod member;
 pub mod protection;
 pub mod secret;
+pub mod sync;
 pub mod update;
 
 pub use add::item;